renamed file; added wrapArgInQuotes function
